Renamed activateMotor to match new terminology
